+++ luapress/config.lua
@@ -1,3 +1,3 @@
 return {
-    version = '2.1.1'
+    version = '3.0'
 }