+++ tests/blog/cleanup_test.sh
@@ -1,0 +1,2 @@
+# Remove config.lua in this test to ensure init config generation works
+rm -rf build/ config.lua